+++ tests/training/trainer_test.py
@@ -18,7 +18,9 @@ from allennlp.common.testing import AllenNlpTestCase, requires_gpu, requires_mul
 from allennlp.data import Vocabulary, Instance
 from allennlp.data.fields import TensorField
 from allennlp.data.data_loaders import MultiProcessDataLoader, SimpleDataLoader, TensorDict
-from allennlp.data.dataset_readers import SequenceTaggingDatasetReader
+from allennlp.data.dataset_readers import SequenceTaggingDatasetReader, DatasetReader
+from allennlp.data.token_indexers import SingleIdTokenIndexer
+from allennlp.data import Token, Instance
 from allennlp.models.model import Model
 from allennlp.models.simple_tagger import SimpleTagger
 from allennlp.training import (
@@ -28,15 +30,72 @@ from allennlp.training import (
     TrackEpochCallback,
     TensorBoardCallback,
     SanityCheckCallback,
+    ConsoleLoggerCallback,
 )
 from allennlp.training.learning_rate_schedulers import CosineWithRestarts
 from allennlp.training.learning_rate_schedulers import ExponentialLearningRateScheduler
 from allennlp.training.momentum_schedulers import MomentumScheduler
 from allennlp.training.moving_average import ExponentialMovingAverage
+from allennlp.data.fields import (
+    TextField,
+    IndexField,
+    MetadataField,
+    LabelField,
+    MultiLabelField,
+    SpanField,
+    FlagField,
+    AdjacencyField,
+    TensorField,
+)
 from allennlp.training.optimizers import Optimizer
 from allennlp.common.testing.sanity_check_test import (
     FakeModelForTestingNormalizationBiasVerification,
 )
+
+
+class FakeDatasetReader(DatasetReader):
+    def __init__(self, total_instances, batch_size):
+        super().__init__()
+        self.total_instances = total_instances
+        self.batch_size = batch_size
+
+    def _read(self, file_path):
+        for i in range(self.total_instances):
+            yield self.text_to_instance(i, "label")
+
+    def text_to_instance(self, index: int, field_type: str):  # type: ignore
+        field = TextField(
+            [Token(t) for t in ["The", "number", "is", str(index), "."]],
+            token_indexers={"words": SingleIdTokenIndexer("words")},
+        )
+
+        return Instance(
+            {
+                "text": field,
+                "label": LabelField(index, skip_indexing=True),
+                "flag": FlagField(23),
+                "index": IndexField(index % self.batch_size, field),
+                "metadata": MetadataField({"some_key": "This will not be logged as a histogram."}),
+                "adjacency": AdjacencyField([(0, 1), (1, 2)], field),
+                "multilabel": MultiLabelField(["l1", "l2"]),
+                "span": SpanField(2, 3, field),
+                "tensor": TensorField(torch.randn(2, 3)),
+            }
+        )
+
+
+class FakeModel(Model):
+    def __init__(self, vocab):
+        super().__init__(vocab)
+        self.lin = torch.nn.Linear(1, 2)
+        self.loss_fn = torch.nn.MSELoss()
+
+    def forward(self, **kwargs):
+        out = kwargs["label"].sum().unsqueeze(-1)
+        out = out.type(torch.FloatTensor)
+        out = self.lin(out)
+        loss = out.sum()
+        return {"loss": loss}
 
 
 class TrainerTestBase(AllenNlpTestCase):
@@ -645,7 +704,7 @@ class TestTrainer(TrainerTestBase):
             serialization_dir=self.TEST_DIR,
             callbacks=[
                 TensorBoardCallback.from_params(
-                    Params({"tensorboard_writer": {"histogram_interval": 2}}),
+                    Params({"tensorboard_writer": {"distribution_interval": 2}}),
                     serialization_dir=self.TEST_DIR,
                 )
             ],
@@ -774,7 +833,6 @@ class TestTrainer(TrainerTestBase):
             callbacks=[SanityCheckCallback(serialization_dir=self.TEST_DIR)],
         )
         trainer.train()
-        assert False
 
     def test_trainer_saves_models_at_specified_interval(self):
         data_loader = SimpleDataLoader(self.instances, 4)
@@ -1092,6 +1150,72 @@ class TestTrainer(TrainerTestBase):
         metrics = trainer.train()
 
         assert metrics["training_loss"] == float(sum(trainer.batch_losses) / batches_per_epoch)
+
+    def test_trainer_can_log_batch_inputs(self):
+        total_instances = 1000
+        batch_size = 25
+
+        reader = FakeDatasetReader(total_instances, batch_size)
+        data_loader = SimpleDataLoader.from_dataset_reader(
+            reader, "fake_path", batch_size=batch_size
+        )
+        instances = list(data_loader.iter_instances())
+        vocab = Vocabulary.from_instances(instances)
+        data_loader.index_with(vocab)
+        model = FakeModel(vocab)
+        optimizer = torch.optim.SGD(model.parameters(), 0.01, momentum=0.9)
+
+        trainer = GradientDescentTrainer(
+            model,
+            optimizer,
+            data_loader,
+            num_epochs=2,
+            serialization_dir=self.TEST_DIR,
+            callbacks=[
+                TensorBoardCallback.from_params(
+                    Params(
+                        {
+                            "tensorboard_writer": {
+                                "distribution_interval": 2,
+                                "should_log_inputs": True,
+                            }
+                        }
+                    ),
+                    serialization_dir=self.TEST_DIR,
+                )
+            ],
+        )
+        trainer.train()
+
+    def test_console_log_callback(self):
+
+        total_instances = 1000
+        batch_size = 25
+
+        reader = FakeDatasetReader(total_instances, batch_size)
+        data_loader = SimpleDataLoader.from_dataset_reader(
+            reader, "fake_path", batch_size=batch_size
+        )
+        instances = list(data_loader.iter_instances())
+        vocab = Vocabulary.from_instances(instances)
+        data_loader.index_with(vocab)
+        model = FakeModel(vocab)
+        optimizer = torch.optim.SGD(model.parameters(), 0.01, momentum=0.9)
+
+        trainer = GradientDescentTrainer(
+            model,
+            optimizer,
+            data_loader,
+            num_epochs=3,
+            serialization_dir=self.TEST_DIR,
+            callbacks=[
+                ConsoleLoggerCallback.from_params(
+                    Params({"should_log_inputs": True}),
+                    serialization_dir=self.TEST_DIR,
+                )
+            ],
+        )
+        trainer.train()
 
 
 @requires_gpu