+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+
+- `histogram_interval` parameter is now deprecated in `TensorboardWriter`, please use `distribution_interval` instead.
+- Memory usage is not logged in tensorboard during training now. `ConsoleLoggerCallback` should be used instead.
+
+### Added
+
+- Added `LogWriter` class. `TensorBoardWriter` now inherits from `LogWriter`.
+- Added `LogCallback` and `ConsoleLoggerCallback` classes. `TensorBoardCallback` inherits from `LogCallback`. 
+
 ## [v2.1.0](https://github.com/allenai/allennlp/releases/tag/v2.1.0) - 2021-02-24
 
 ### Changed
@@ -18,6 +30,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added a way to specify extra parameters to the predictor in an `allennlp predict` call.
 - Added a way to initialize a `Vocabulary` from transformers models.
 - Added `NormalizationBiasVerification` and `SanityCheckCallback` for model sanity checks.
+- Added a way to initialize a `Vocabulary` from transformers models. 
 - Added the ability to use `Predictors` with multitask models through the new `MultiTaskPredictor`.
 - Added an example for fields of type `ListField[TextField]` to `apply_token_indexers` API docs.
 - Added `text_key` and `label_key` parameters to `TextClassificationJsonReader` class.